Move search commands (like `/` and `*`) into their own file
