+++ src/actions/include-all.ts
@@ -8,6 +8,7 @@ import './commands/insert';
 import './commands/replace';
 import './commands/actions';
 import './commands/commandLine';
+import './commands/search';
 
 // plugin
 import './plugins/camelCaseMotion';
